Implementation of debug uart added.
Implementation of debug uart added: printf uses dma in order to achieve non-blocking transmission of data. debug uart uses a custom fifo implementation. Data is pushed into the fifo when dma transfer is in progress and data pulled from the fifo when the uart tx is completed. fifo implementation is tested. printf is partially tested.